Messages stored in state
Done more the react way. Some improvements in how messages are stored would be helpful.

TODO: Make a single array of arrays hold all the messages. Right now, we have a send to all array, and a send to one player array. There might be the case where we want to send some data to one player, and other data to another player.